L5 - input done with unordered_map

+++ MyGameEngine/Bitmask.cpp
@@ -10,12 +10,12 @@ uint32_t Bitmask::GetMask() const {
     return bits;
 }
 
-bool Bitmask::GetBit(int pos) const {
+bool Bitmask::GetBit(size_t pos) const {
     return (bits & (1 << pos)) != 0;
 }
 
 // A simple helper method that calls set or clear bit
-void Bitmask::SetBit(int pos, bool on) {
+void Bitmask::SetBit(size_t pos, bool on) {
     if (on) {
         SetBit(pos);
     }
@@ -24,11 +24,11 @@ void Bitmask::SetBit(int pos, bool on) {
     }
 }
 
-void Bitmask::SetBit(int pos) {
+void Bitmask::SetBit(size_t pos) {
     bits = bits | 1 << pos;
 }
 
-void Bitmask::ClearBit(int pos) {
+void Bitmask::ClearBit(size_t pos) {
     bits = bits & ~(1 << pos);
 }
